Add branch instruction to empty directory interface for the case of directory not exists.

+++ apyscript/file/file_util.py
@@ -15,5 +15,6 @@ def empty_directory(directory_path: str) -> None:
         Directory path to empty. This folder itself will not be
         removed.
     """
-    shutil.rmtree(directory_path, ignore_errors=True)
+    if os.path.isdir(directory_path):
+        shutil.rmtree(directory_path, ignore_errors=True)
     os.makedirs(directory_path, exist_ok=True)
